+++ src/backend/postgres/foreign_key.rs
@@ -17,7 +17,14 @@ impl ForeignKeyBuilder for PostgresQueryBuilder {
 
         write!(sql, "DROP CONSTRAINT ").unwrap();
         if let Some(name) = &drop.foreign_key.name {
-            write!(sql, "\"{name}\"").unwrap();
+            write!(
+                sql,
+                "{}{}{}",
+                self.quote().left(),
+                name,
+                self.quote().right()
+            )
+            .unwrap();
         }
     }
 
@@ -41,7 +48,14 @@ impl ForeignKeyBuilder for PostgresQueryBuilder {
 
         if let Some(name) = &create.foreign_key.name {
             write!(sql, "CONSTRAINT ").unwrap();
-            write!(sql, "\"{name}\" ").unwrap();
+            write!(
+                sql,
+                "{}{}{} ",
+                self.quote().left(),
+                name,
+                self.quote().right()
+            )
+            .unwrap();
         }
 
         write!(sql, "FOREIGN KEY (").unwrap();