+++ src/index/create.rs
@@ -153,7 +153,10 @@ impl IndexCreateStatement {
     }
 
     /// Set index name
-    pub fn name(&mut self, name: &str) -> &mut Self {
+    pub fn name<T>(&mut self, name: T) -> &mut Self
+    where
+        T: Into<String>,
+    {
         self.index.name(name);
         self
     }