+++ src/query/with.rs
@@ -139,7 +139,7 @@ impl CommonTableExpression {
     }
 
     fn set_table_name_from_select(&mut self, iden: &DynIden) {
-        self.table_name = Some(Alias::new(&format!("cte_{}", iden.to_string())).into_iden())
+        self.table_name = Some(Alias::new(format!("cte_{}", iden.to_string())).into_iden())
     }
 
     /// Set up the columns of the CTE to match the given [SelectStatement] selected columns.
@@ -162,15 +162,11 @@ impl CommonTableExpression {
                         SimpleExpr::Column(column) => match column {
                             ColumnRef::Column(iden) => Some(iden.clone()),
                             ColumnRef::TableColumn(table, column) => Some(
-                                Alias::new(&format!(
-                                    "{}_{}",
-                                    table.to_string(),
-                                    column.to_string()
-                                ))
-                                .into_iden(),
+                                Alias::new(format!("{}_{}", table.to_string(), column.to_string()))
+                                    .into_iden(),
                             ),
                             ColumnRef::SchemaTableColumn(schema, table, column) => Some(
-                                Alias::new(&format!(
+                                Alias::new(format!(
                                     "{}_{}_{}",
                                     schema.to_string(),
                                     table.to_string(),