+++ src/token.rs
@@ -293,6 +293,7 @@ impl std::fmt::Display for Token {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use pretty_assertions::assert_eq;
 
     #[test]
     fn test_0() {
@@ -382,7 +383,7 @@ mod tests {
         let string = r#""a\"bc""#;
         let tokenizer = Tokenizer::new(string);
         let tokens: Vec<Token> = tokenizer.iter().collect();
-        assert_eq!(tokens, vec![Token::Quoted("\"a\\\"bc\"".to_string()),]);
+        assert_eq!(tokens, vec![Token::Quoted("\"a\\\"bc\"".to_string())]);
         assert_eq!(
             string,
             tokens.iter().map(|x| x.to_string()).collect::<String>()
@@ -394,7 +395,7 @@ mod tests {
         let string = "abc123";
         let tokenizer = Tokenizer::new(string);
         let tokens: Vec<Token> = tokenizer.iter().collect();
-        assert_eq!(tokens, vec![Token::Unquoted(string.to_string()),]);
+        assert_eq!(tokens, vec![Token::Unquoted(string.to_string())]);
         assert_eq!(
             string,
             tokens.iter().map(|x| x.to_string()).collect::<String>()
@@ -602,7 +603,7 @@ mod tests {
         let string = "abc_$123";
         let tokenizer = Tokenizer::new(string);
         let tokens: Vec<Token> = tokenizer.iter().collect();
-        assert_eq!(tokens, vec![Token::Unquoted(string.to_string()),]);
+        assert_eq!(tokens, vec![Token::Unquoted(string.to_string())]);
         assert_eq!(
             string,
             tokens.iter().map(|x| x.to_string()).collect::<String>()