+++ tests/sqlite/table.rs
@@ -1,4 +1,5 @@
 use super::*;
+use pretty_assertions::assert_eq;
 
 #[test]
 fn create_1() {
@@ -403,6 +404,7 @@ fn alter_6() {
 }
 
 #[test]
+#[should_panic(expected = "Sqlite doesn't support multiple alter options")]
 fn alter_7() {
     let _ = Table::alter()
         .table(Font::Table)